팀, 유저 선택 수정 완료 search.js 스크립트 파일 분리 완료
