+++ src/main/resources/templates/common/head.html
@@ -11,3 +11,6 @@
 <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
 <script th:src="@{/js/style.js}" defer></script>
 
+<script>
+    const delete_member_list = [];
+</script>
