+++ src/main/resources/templates/member/detail.html
@@ -109,5 +109,5 @@
 </main>
 </body>
 <script th:src="@{/js/member/detail.js}"></script>
-<script th:src="@{/js/search.js}"></script>
+<script th:src="@{/js/team/search.js}"></script>
 </html>
