+++ src/main/resources/templates/project/detail.html
@@ -96,5 +96,6 @@
 </main>
 </body>
 <script th:src="@{/js/project/detail.js}"></script>
-<script th:src="@{/js/search.js}"></script>
+<script th:src="@{/js/member/search.js}"></script>
+<script th:src="@{/js/team/search.js}"></script>
 </html>
